refactor: split NoahTools nitro module into smaller files
- Split iOS NoahTools.swift into separate files for HTTP, logging, and backup functionality
- Split Android NoahTools.kt into separate Kotlin objects for HTTP, logging, and backup
- Main NoahTools classes now delegate to helper extensions/objects
- Improves code organization and maintainability
